requirements.txt
h11~=0.14.0
pip~=23.2.1
wheel~=0.41.2
requests~=2.32.3
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.7
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
jiter~=0.8.0
distro~=1.9.0
openai~=1.57.0
pydantic~=2.9.2
anyio~=4.7.0
numpy~=2.0.2
pandas~=2.2.3
certifi~=2024.8.30
Deprecated~=1.2.15
setuptools~=68.2.0
uvicorn~=0.32.1
fastapi~=0.115.6
python-dotenv~=1.0.1
dash